fix: manual clippy fixes (loop iterators, casts, strip_prefix, char find)
Hand-applied structural clippy fixes across 16 files. No imports,
variables, fields, or `mut` qualifiers removed — wired and unwired
code preserved verbatim.

Categories applied:
- Loop variable used to index → iterator form (fingerprint, high_heel,
  proprioception, context_chain, sparql_parse, nars_engine, api,
  audio/node, sensorium, clam_neighborhood, orchestration_mode)
- Manual char comparison → array literal in find()/split() (literal_graph,
  exploration)
- Manually stripping a prefix → strip_prefix() (literal_graph)
- Redundant cast (u32→u32, usize→usize) (graph_router, query, chat_bundle,
  context_chain)

233 contract + 170 planner tests pass. Workspace compiles clean.

+++ crates/lance-graph-planner/src/strategy/chat_bundle.rs
@@ -44,8 +44,8 @@ impl AutocompleteCache {
         self.turn_count += 1;
 
         // 1. Update user model
-        let row = (self.turn_count as usize % 64) as usize;
-        let col = (self.turn_count as usize / 64 % 64) as usize;
+        let row = self.turn_count as usize % 64;
+        let col = self.turn_count as usize / 64 % 64;
         self.triple.on_user_input(message, row, col);
 
         // 2. Evaluate all 4096 heads with DK-appropriate tension
@@ -86,8 +86,8 @@ impl AutocompleteCache {
 
     /// After LLM generates, update models.
     pub fn on_self_output(&mut self, output: &HeadPrint) {
-        let row = (self.turn_count as usize % 64) as usize;
-        let col = (self.turn_count as usize / 64 % 64) as usize;
+        let row = self.turn_count as usize % 64;
+        let col = self.turn_count as usize / 64 % 64;
         self.triple.on_self_output(output, row, col);
     }
